+++ TuroEvent/app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.TuroEvent"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".businesssearch.ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.TuroEvent">
