+++ TuroEvent/app/src/main/java/com/ameer/turoevent/businesssearch/data/BusinessSearchApiResponse.kt
@@ -3,7 +3,7 @@ package com.example.example
 import com.google.gson.annotations.SerializedName
 
 
-data class ExampleJson2KtKotlin (
+data class BusinessSearchApiResponse (
 
   @SerializedName("businesses" ) var businesses : ArrayList<Businesses> = arrayListOf(),
   @SerializedName("total"      ) var total      : Int?                  = null,
